move addCurrentFile and addIgnoreCurrentFile to main.ts
